Working on mining/surveys

+++ src-tauri/src/models/ship/cooldown.rs
@@ -5,8 +5,8 @@ pub struct Cooldown {
   #[serde(rename = "shipSymbol")]
   pub ship_symbol: String,
   #[serde(rename = "totalSeconds")]
-  pub total_seconds: u64,
+  pub total_seconds: i32,
   #[serde(rename = "remainingSeconds")]
-  pub remaining_seconds: u64,
+  pub remaining_seconds: i32,
   pub expiration: Option<String>
 }
